+++ app/system/controllers/search_user.php
@@ -2,10 +2,11 @@
 
 $query = isset($_GET['query']) ? trim($_GET['query']) : '';
 
-if (!empty($query)) {    
-    $results = ORM::for_table('tbl_customers')
-        ->where_like('username', "%$query%")
-        ->find_many();
+if (!empty($query)) {
+    $rq = ORM::for_table('tbl_customers')
+        ->where_like('username', "%$query%");
+    AgentScope::filterCustomersQuery($rq, $admin);
+    $results = $rq->find_many();
 
     if ($results) {
         echo '<ul>';